Remove Contact & improve structure of code in order-confirm.comp
- Contact interface is removed. It only contains two properties, and which
are retrived from the Contact object to build the Order object. The
Contact object is never reused.

- The "cart" as a map object contains the references to the Book objects
received from the backend as well as other extra information (e.g., amount).
Nonetheless, when building an Order object to send to the backend server, only
the book id for each book is needed and only List is accepted in backend.

To improve structure of code and readability, the previous code that converts
the cart to a List is moved to checkout.service as a new method, as the cart
should be managed by this service rather than the component.
