--- frontend/frontend/src/app/model/contact.ts
@@ -1,4 +1,0 @@
-export interface Contact {
-  firstName: string;
-  lastName: string;
-}
