+++ frontend/frontend/src/app/order-confirm/order-confirm.component.html
@@ -75,12 +75,11 @@
         <div class="form-group">
             <label for="firstName"><span style="color: brown;">*</span> First Name</label>
             <input type="text" class="form-control" placeholder="Enter first name" id="firstName"
-                [(ngModel)]="contact.firstName">
+                [(ngModel)]="firstName">
         </div>
         <div class="form-group">
             <label for="lastName"><span style="color: brown;">*</span> Last Name</label>
-            <input type="text" class="form-control" placeholder="Enter last name" id="lastName"
-                [(ngModel)]="contact.lastName">
+            <input type="text" class="form-control" placeholder="Enter last name" id="lastName" [(ngModel)]="lastName">
         </div>
     </div>
     <div class=" container p-4 m-4 text-right">
